requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.4
groq>=0.9.0
python-dotenv>=1.0.1
requests>=2.31.0
streamlit>=1.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
